Separate command to multiple files.
