+++ slider/src/app/commands.h
@@ -11,20 +11,6 @@ using namespace Core;
 
 namespace Slider
 {
-    class MaxSpeedCommand : public MenuCommand
-    {
-    public:
-        virtual void Print(Core::Display* display) const override;
-        virtual void Invoke(MenuCommandButton command) override;
-    };
-
-    class SpeedCurveCommand : public MenuCommand
-    {
-    public:
-        virtual void Print(Core::Display* display) const override;
-        virtual void Invoke(MenuCommandButton command) override;
-    };
-
     class BrainAddressCommand : public MenuCommand
     {
     public:
