+++ slider/src/app/simulator/simulatorApp.cpp
@@ -5,6 +5,9 @@
 #include "src/core/perf/perf.h"
 #include "src/network/wifiComponent.h"
 
+#include "src/commands/speedCommand.h"
+#include "src/commands/curveCommand.h"
+
 #include <esp32-hal-timer.h>
 
 using namespace Input;
