+++ slider/src/commands/speedCommand.cpp
@@ -1,0 +1,37 @@
+#include "speedCommand.h"
+#include "src/app/settings.h"
+#include "src/core/utils/mathUtils.h"
+
+using namespace Core;
+using namespace Slider;
+
+const char* GetDesc(Settings::Speed speed)
+{
+    switch (speed)
+    {
+        case Settings::SpeedSlow:
+            return "Pas vite";
+        case Settings::SpeedMedium:
+            return "Moyen vite";
+        case Settings::SpeedHigh:
+            return "Fuckin vite";
+        default:
+            return nullptr;
+    };
+}
+
+void MaxSpeedCommand::Print(Display* display) const
+{
+    PrintTitle(display, "Vitesse max.");
+    auto desc = GetDesc(Settings::GetInstance().GetSpeed());
+    PrintDescription(display, desc);
+}
+
+void MaxSpeedCommand::Invoke(MenuCommandButton command)
+{
+    const auto speed = Settings::GetInstance().GetSpeed();
+    if (command == MenuCommand::ButtonLeft)
+        Settings::GetInstance().SetSpeed(GetPreviousEnumValue(speed));
+    else if (command == MenuCommand::ButtonRight)
+        Settings::GetInstance().SetSpeed(GetNextEnumValue(speed));
+}
